LayoutTests/storage/indexeddb/mozilla/add-twice-failure-private-expected.txt

Test IndexedDB behavior adding the same property twice

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('foo');
key = 10
request = objectStore.add({}, key);
PASS request.result is key
request = objectStore.add({}, key);
PASS event.target.error.name is 'ConstraintError'
PASS successfullyParsed is true

TEST COMPLETE
